+++ Sources/FeatherDatabase/PostgresDatabase/PostgresQuery.swift
@@ -1,0 +1,7 @@
+import PostgresNIO
+
+extension PostgresQuery: DatabaseQuery {
+    public typealias Bindings = PostgresBindings
+
+    public var bindings: PostgresBindings { binds }
+}
