+++ Sources/FeatherDatabase/SQLiteDatabase/SQLiteRow+DatabaseRow.swift
@@ -1,0 +1,103 @@
+import SQLiteNIO
+
+extension SQLiteRow: DatabaseRow {
+
+    struct SingleValueDecoder: Decoder, SingleValueDecodingContainer {
+
+        var codingPath: [any CodingKey] { [] }
+        var userInfo: [CodingUserInfoKey: Any] { [:] }
+
+        let data: SQLiteData
+
+        func container<Key: CodingKey>(
+            keyedBy: Key.Type
+        ) throws(DecodingError) -> KeyedDecodingContainer<Key> {
+            throw DecodingError.typeMismatch(
+                KeyedDecodingContainer<Key>.self,
+                .init(
+                    codingPath: codingPath,
+                    debugDescription: "Keyed decoding is not supported."
+                )
+            )
+        }
+
+        func unkeyedContainer() throws(DecodingError)
+            -> any UnkeyedDecodingContainer
+        {
+            throw DecodingError.typeMismatch(
+                (any UnkeyedDecodingContainer).self,
+                .init(
+                    codingPath: codingPath,
+                    debugDescription: "Unkeyed decoding is not supported."
+                )
+            )
+        }
+
+        func singleValueContainer() throws(DecodingError)
+            -> any SingleValueDecodingContainer
+        {
+            self
+        }
+
+        func decodeNil() -> Bool {
+            data.isNull
+        }
+
+        func decode<T: Decodable>(
+            _ type: T.Type
+        ) throws(DecodingError) -> T {
+            if let convertible = type as? any SQLiteDataConvertible.Type {
+                guard let value = convertible.init(sqliteData: data) else {
+                    throw DecodingError.typeMismatch(
+                        T.self,
+                        .init(
+                            codingPath: codingPath,
+                            debugDescription:
+                                "Could not convert data to \(T.self): \(data)."
+                        )
+                    )
+                }
+                return value as! T
+            }
+
+            throw DecodingError.typeMismatch(
+                T.self,
+                .init(
+                    codingPath: codingPath,
+                    debugDescription:
+                        "Data is not convertible to \(T.self): \(data)."
+                )
+            )
+        }
+    }
+
+    public func decode<T: Decodable>(
+        column: String,
+        as type: T.Type
+    ) throws(DecodingError) -> T {
+        guard let data = self.column(column) else {
+            throw .dataCorrupted(
+                .init(
+                    codingPath: [],
+                    debugDescription: "Missing data for column \(column)."
+                )
+            )
+        }
+        do {
+            return try T(from: SingleValueDecoder(data: data))
+        }
+        catch let error as DecodingError {
+            throw error
+        }
+        catch {
+            throw DecodingError.typeMismatch(
+                T.self,
+                .init(
+                    codingPath: [],
+                    debugDescription:
+                        "Data is not convertible to \(T.self): \(data)."
+                )
+            )
+        }
+    }
+}
