+++ Tests/FeatherDatabaseTests/SQLiteDatabaseTestSuite.swift
@@ -16,9 +16,7 @@ import Testing
 @Suite
 struct SQLiteDatabaseTestSuite {
 
-    private func getTestDatabase() async throws
-        -> FeatherDatabase.SQLiteDatabase
-    {
+    private func getTestDatabaseClient() async throws -> SQLiteDatabaseClient {
         var logger = Logger(label: "test")
         logger.logLevel = .debug
 
@@ -27,7 +25,7 @@ struct SQLiteDatabaseTestSuite {
             logger: logger
         )
 
-        return SQLiteDatabase(
+        return .init(
             connection: connection,
             logger: logger
         )
@@ -35,8 +33,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func foreignKeySupport() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: "PRAGMA foreign_keys = ON"
@@ -54,8 +52,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func tableCreation() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -85,8 +83,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func tableInsert() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -130,8 +128,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func rowDecoding() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -185,8 +183,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func queryEncoding() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         let tableName = "foo"
         let idColumn = "id"
@@ -237,8 +235,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func unsafeSQLBindings() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -277,8 +275,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func optionalStringInterpolationNil() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -315,8 +313,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func sqliteDataInterpolation() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -353,8 +351,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func resultSequenceIterator() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -411,8 +409,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func transactionSuccess() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -450,8 +448,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func transactionFailurePropagates() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -509,8 +507,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func doubleRoundTrip() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -548,8 +546,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func missingColumnThrows() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -594,8 +592,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func typeMismatchThrows() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         try await database.execute(
             query: #"""
@@ -640,8 +638,8 @@ struct SQLiteDatabaseTestSuite {
 
     @Test
     func versionCheck() async throws {
-        let database = try await getTestDatabase()
-        defer { Task { try await database.shutdown() } }
+        let database = try await getTestDatabaseClient()
+        // defer { Task { try await database.shutdown() } }
 
         let result = try await database.execute(
             query: #"""
